run: bitters install and remove main.scss
Bitters is a handy reorganised set of default stylesheets that
saves us time setting up our styles.

It covers most of the content in main.scss so I've removed that as well
with this commit.

The print styles in main.scss aren't covered now.
We haven't confirmed there's a user need for that yet
so I'm no bothered by losing them for now.