Enable the translate to work from any path
Currently the translation process will take you back to the home page
unless you've on doctor search results. If you're on the doctors
index with no query, you are returned to the home page.

This makes the process more flexible so it will take you back to any
page that you wanted to translate. This is required before enabling
translations of a /about page or something else.